TextEditor
This project is about creating the basic TextEditor and implementing some basic operation like open file, save file, close window, open window etc which is similar to somewhat Notepad. Text Editor is a .txt File Editor in which we can Edit the File it has some basic functionality which needs to edit the file.